\name{euclidean.distance}
\alias{euclidean.distance}
\title{Euclidean distance}
\description{Calculates the Euclidean distance between two vectors}
\usage{
euclidean.distance(x,y);
}
\arguments{
  \item{x}{A vector of coordinates}
  \item{y}{A vector of coordinates}
}
\value{The euclidean distance}
\author{Paul C. Boutros}
\examples{
euclidean.distance( x = c(0,0,0), y = c(1,1,1) );
}
\keyword{manip}
